backend/requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==5.3.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
dotenv==0.9.9
fastapi==0.121.2
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.11.0
numpy==1.26.4
openai==1.4.0
pandas==2.2.0
pillow==10.2.0
protobuf==4.25.3
pyarrow==15.0.0
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rpds-py==0.28.0
six==1.17.0
smmap==3.0.1
sniffio==1.3.0
starlette==0.49.3
streamlit==1.31.0
tenacity==8.4.0
toml==0.10.2
tornado==6.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
urllib3==2.2.0
uvicorn==0.27.0
google-generativeai==0.3.0